# bank-lending-frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation for Spring Security",
    "contact": {
      "name": "Timur Sharf",
      "url": "https://github.com/xzcPozer",
      "email": "timurkasharf@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/credit-query/edit/solvency-verify/{userId}": {
      "put": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "solvencyVerifyPdf",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/edit/query-verify/{userId}": {
      "put": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "creditQueryVerifyPdf",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/edit/employment-verify/{userId}": {
      "put": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "employmentVerifyPdf",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/change/solvency/ip/tax-agent-info/{userId}": {
      "put": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "changeSolvencyIpTaxAgent",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditQueryTaxAgentIpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/change/solvency/ip/financial-situation/{userId}": {
      "put": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "changeSolvencyIpFinancialSituation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditQueryFinancialSituationIpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/change/solvency/hired-worker/tax-agent-info/{userId}": {
      "put": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "changeSolvencyHiredWorkerTaxAgent",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditQueryTaxAgentHiredWorkerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/change/solvency/hired-worker/financial-situation/{userId}": {
      "put": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "changeSolvencyHiredWorkerFinancialSituation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditQueryFinancialSituationHiredWorkerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/credit-query/send/refuse/{userId}": {
      "post": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "sendRefuseRequest",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "description"
                ],
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/send/for-revision/{userId}": {
      "post": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "sendForRevisionRequest",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "description"
                ],
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/generate-query/{userId}": {
      "post": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "createCreditQueryPdf",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/confirmation-solvency/{userId}": {
      "post": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "confirmationOfSolvencyPdf",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/confirmation-employment/{userId}": {
      "post": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "confirmationOfEmploymentPdf",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/credit-query/get/solvency/{userId}": {
      "get": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "getSolvencyInformation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/get/solvency/upload/{userId}": {
      "get": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "getSolvencyUploadInformation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/get/payment/{userId}": {
      "get": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "getPaymentInformation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/get/payment/upload/{userId}": {
      "get": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "getPaymentUploadInformation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/get/credit-request/{userId}": {
      "get": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "getCreditRequestInformation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/all-client": {
      "get": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "getPaymentInformation_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCreditQueryClientResponse"
                }
              }
            }
          }
        }
      }
    },
    "/credit-query/all-client/by/lastname": {
      "get": {
        "tags": [
          "CreditQuery"
        ],
        "operationId": "getAllClientPaymentInformationByLastname",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "lastname",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCreditQueryClientResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreditQueryTaxAgentIpRequest": {
        "required": [
          "inn",
          "name",
          "ogrnip"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "inn": {
            "type": "string"
          },
          "ogrnip": {
            "type": "string"
          }
        }
      },
      "CreditQueryFinancialSituationIpRequest": {
        "required": [
          "monthlyPayment",
          "totalIncome"
        ],
        "type": "object",
        "properties": {
          "totalIncome": {
            "type": "string"
          },
          "monthlyPayment": {
            "type": "string"
          }
        }
      },
      "CreditQueryTaxAgentHiredWorkerRequest": {
        "required": [
          "inn",
          "kpp",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "inn": {
            "type": "string"
          },
          "kpp": {
            "type": "string"
          }
        }
      },
      "CreditQueryFinancialSituationHiredWorkerRequest": {
        "required": [
          "monthlyPayment",
          "taxCalculationAmount",
          "totalIncome"
        ],
        "type": "object",
        "properties": {
          "totalIncome": {
            "type": "string"
          },
          "taxCalculationAmount": {
            "type": "string"
          },
          "monthlyPayment": {
            "type": "string"
          }
        }
      },
...
      "CreditQueryClientResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "surName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "inn": {
            "type": "string"
          },
          "kpp": {
            "type": "string"
          },
          "totalIncome": {
            "type": "string"
          },
          "taxCalculationAmount": {
            "type": "string"
          },
          "monthlyPayment": {
            "type": "string"
          },
          "immovableProperty": {
            "type": "boolean"
          },
          "movableProperty": {
            "type": "boolean"
          }
        }
      },
      "PageResponseCreditQueryClientResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditQueryClientResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth with bearer token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}